# autopilot-api/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/social-accounts/connect": {
      "post": {
        "operationId": "SocialAccountsController_connect",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SocialAccountsCreateDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Connect a social account manually (tokens/credentials)",
        "tags": [
          "Social Accounts"
        ]
      }
    },
    "/api/v1/social-accounts/tenant/{tenantId}": {
      "get": {
        "operationId": "SocialAccountsController_findByTenant",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "workspaceId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List connected social accounts for a tenant",
        "tags": [
          "Social Accounts"
        ]
      }
    },
    "/api/v1/social-accounts/me": {
      "get": {
        "operationId": "SocialAccountsController_getMyAccounts",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get connected social accounts for the current user",
        "tags": [
          "Social Accounts"
        ]
      }
    },
    "/api/v1/social-accounts/oauth/{platform}/authorize": {
      "get": {
        "operationId": "SocialAccountsController_startOAuth",
        "parameters": [
          {
            "name": "platform",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tenantId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "returnUrl",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "workspaceId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Start OAuth flow to connect a platform for a tenant",
        "tags": [
          "Social Accounts"
        ]
      }
    },
    "/api/v1/social-accounts/oauth/{platform}/callback": {
      "get": {
        "operationId": "SocialAccountsController_oauthCallback",
        "parameters": [
          {
            "name": "platform",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "code",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "state",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "error",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "error_description",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "OAuth callback — connects account and redirects to returnUrl",
        "tags": [
          "Social Accounts"
        ]
      }
    },
    "/api/v1/social-accounts/facebook/setup": {
      "get": {
        "operationId": "SocialAccountsController_getFacebookSetup",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Preview Facebook Pages from OAuth setup token",
        "tags": [
          "Social Accounts"
        ]
      }
    },
    "/api/v1/social-accounts/facebook/finalize": {
      "post": {
        "operationId": "SocialAccountsController_finalizeFacebook",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FacebookFinalizeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Finalize Facebook connect after selecting a Page",
        "tags": [
          "Social Accounts"
        ]
      }
    },
    "/api/v1/social-accounts/youtube/setup": {
      "get": {
        "operationId": "SocialAccountsController_getYoutubeSetup",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Preview YouTube channels from OAuth setup token",
        "tags": [
          "Social Accounts"
        ]
      }
    },
    "/api/v1/social-accounts/youtube/finalize": {
      "post": {
        "operationId": "SocialAccountsController_finalizeYoutube",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/YoutubeFinalizeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Finalize YouTube connect after selecting a channel",
        "tags": [
          "Social Accounts"
        ]
      }
    },
    "/api/v1/social-accounts/whatsapp/enable-platform": {
      "post": {
        "operationId": "SocialAccountsController_enablePlatformWhatsapp",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Enable platform-managed WhatsApp for a workspace (no client Meta setup)",
        "tags": [
          "Social Accounts"
        ]
      }
    },
    "/api/v1/social-accounts/whatsapp/setup-from-meta": {
      "post": {
        "operationId": "SocialAccountsController_setupWhatsappFromMeta",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Try WhatsApp setup using an existing Facebook connection (skips Meta login when scopes allow)",
        "tags": [
          "Social Accounts"
        ]
      }
    },
    "/api/v1/social-accounts/whatsapp/setup": {
      "get": {
        "operationId": "SocialAccountsController_getWhatsAppSetup",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Preview WhatsApp phone numbers from OAuth setup token",
        "tags": [
          "Social Accounts"
        ]
      }
    },
    "/api/v1/social-accounts/whatsapp/finalize": {
      "post": {
        "operationId": "SocialAccountsController_finalizeWhatsApp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WhatsappFinalizeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Finalize WhatsApp connect after selecting a phone number",
        "tags": [
          "Social Accounts"
        ]
      }
    },
    "/api/v1/social-accounts/{id}/disconnect": {
      "post": {
        "operationId": "SocialAccountsController_disconnect",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tenantId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Disconnect a social account",
        "tags": [
          "Social Accounts"
        ]
      }
    },
    "/api/v1/social-accounts/{id}": {
      "delete": {
        "operationId": "SocialAccountsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tenantId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Remove a social account",
        "tags": [
          "Social Accounts"
        ]
      }
    },
...
  },
  "info": {
    "title": "API",
...
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "widget-api-key": {
        "type": "http",
        "in": "header",
        "name": "widget-api-key",
        "scheme": "bearer",
        "bearerFormat": "API Key",
        "description": "Chatbot widget API key (`pk_live_…`). Create keys in AI Chatbot → Embed. Used for `/api/v1/widget/*` routes."
      }
    },
    "schemas": {
...
      "SocialAccountsCreateDto": {
        "type": "object",
        "properties": {}
      },
      "FacebookFinalizeDto": {
        "type": "object",
        "properties": {
          "setupToken": {
            "type": "string",
            "description": "Short-lived setup token from OAuth callback"
          },
          "pageId": {
            "type": "string",
            "description": "Selected Facebook Page ID"
          }
        },
        "required": [
          "setupToken",
          "pageId"
        ]
      },
      "YoutubeFinalizeDto": {
        "type": "object",
        "properties": {
          "setupToken": {
            "type": "string",
            "description": "Short-lived setup token from OAuth callback"
          },
          "channelId": {
            "type": "string",
            "description": "Selected YouTube channel ID"
          }
        },
        "required": [
          "setupToken",
          "channelId"
        ]
      },
      "WhatsappFinalizeDto": {
        "type": "object",
        "properties": {
          "setupToken": {
            "type": "string",
            "description": "Short-lived setup token from OAuth callback"
          },
          "phoneNumberId": {
            "type": "string",
            "description": "Selected WhatsApp phone number ID from Meta"
          }
        },
        "required": [
          "setupToken",
          "phoneNumberId"
        ]
      },
...
    }
  }
}